Crate objc2_cloud_kit

Bindings to the CloudKit framework

See Apple’s docs and the general docs on framework crates for more information.

Structs

CKAccountStatus

credentials in Settings app.

CKApplicationPermissionStatusDeprecated

Apple’s documentation

CKApplicationPermissions

Apple’s documentation

CKContainer

A CKContainer, and its CKDatabases, are the main entry points into the CloudKit framework.

CKRecord

Apple’s documentation

CKShare

Like CKRecords, CKShares can store arbitrary key-value pairs. They are modified and fetched in the same manner. A share, its root record, and its root record’s children records will only appear in a participant’s CKFetchRecordChangesOperation’s results after the share has been accepted by that participant. Clients have access to the share (and optionally the root record) before accepting a share, via the CKShareMetadata object. Note that in order to access a root record before accepting a share, you must run a CKFetchShareMetadataOperation requesting the root record. A CKShare will appear in a CKFetchRecordChangesOperation’s results set whenever the participant list is updated. For that reason, you shouldn’t place heavy key-value pairs in it.

CKShareMetadata

Apple’s documentation

Statics

CKAccountChangedNotification⚠

This local notification is posted when there has been any change to the logged in iCloud account.

CKCurrentUserDefaultName⚠

Stand-in for the current user’s ID; most often used in RecordZoneID->ownerName

CKOwnerDefaultName⚠Deprecated

Apple’s documentation

CKRecordCreationDateKey⚠

For use in queries to match on record properties. Matches record.creationDate. Value is a NSDate

CKRecordCreatorUserRecordIDKey⚠

For use in queries to match on record properties. Matches record.creatorUserRecordID. Value is a CKRecordID

CKRecordLastModifiedUserRecordIDKey⚠

For use in queries to match on record properties. Matches record.lastModifiedUserRecordID. Value is a CKRecordID

CKRecordModificationDateKey⚠

For use in queries to match on record properties. Matches record.modificationDate. Value is a NSDate

CKRecordNameZoneWideShare⚠

A zone-wide CKShare always uses the record name CKRecordNameZoneWideShare.You can use this to fetch the CKSharerecord for the zone with a CKFetchRecordsOperation.

CKRecordParentKey⚠

For use in queries to match on record properties. Matches record.parent

CKRecordRecordIDKey⚠

For use in queries to match on record properties. Matches record.recordID. Value is a CKRecordID

CKRecordShareKey⚠

For use in queries to match on record properties. Matches record.share

CKRecordTypeShare⚠

Apple’s documentation

CKRecordTypeUserRecord⚠

Use this constant for the recordType parameter when fetching User Records.

CKShareThumbnailImageDataKey⚠

Value is a data blob suitable to pass into

CKShareTitleKey⚠

Value is a string. Example for a recipe sharing app: “Pot Roast”

CKShareTypeKey⚠

Value is a string representing a UTI. Example for a recipe sharing app: “com.mycompany.recipe”

Traits

CKRecordKeyValueSetting

Formalizes a protocol for getting and setting keys on a CKRecord. Not intended to be used directly by client code

CKRecordValue

Apple’s documentation

Type Aliases

CKRecordFieldKey

Apple’s documentation

CKRecordType

Apple’s documentation